+++ OnlineRetailer.ProductAPI/Core/ApplicationServices/Services/ProductService.cs
@@ -1,5 +1,6 @@
 ﻿using OnlineRetailer.ProductAPI.Core.DomainServices;
 using OnlineRetailer.ProductAPI.Entities;
+using System;
 using System.Collections.Generic;
 
 namespace OnlineRetailer.ProductAPI.Core.ApplicationServices.Services
@@ -14,29 +15,49 @@ namespace OnlineRetailer.ProductAPI.Core.ApplicationServices.Services
 			_productRepository = productRepository;
 		}
 
-		public Product Add(Product product)
-		{
-			return _productRepository.Add(product);
-		}
-
-		public void Edit(Product product)
-		{
-			_productRepository.Edit(product);
-		}
-
-		public Product Get(int id)
-		{
-			return _productRepository.Get(id);
-		}
-
 		public IEnumerable<Product> GetAll()
 		{
 			return _productRepository.GetAll();
 		}
 
-		public void Remove(int id)
+		public Product Get(int id)
 		{
-			_productRepository.Remove(id);
+			if (_productRepository.Get(id) == null)
+			{
+				throw new NullReferenceException($"Could not find Product with ID: {id}");
+			}
+			return _productRepository.Get(id);
+		}
+
+		public Product Add(Product product)
+		{
+			if (product == null)
+			{
+				throw new NullReferenceException("Product is null");
+			}
+			return _productRepository.Add(product);
+		}
+
+		public Product Edit(Product product)
+		{
+			if (product == null)
+			{
+				throw new NullReferenceException("Product is null");
+			}
+			if (_productRepository.Get(product.Id) == null)
+			{
+				throw new NullReferenceException($"Could not find Product with ID: {product.Id}");
+			}
+			return _productRepository.Edit(product);
+		}
+
+		public Product Remove(int id)
+		{
+			if (_productRepository.Get(id) == null)
+			{
+				throw new NullReferenceException($"Could not find Product with ID: {id}");
+			}
+			return _productRepository.Remove(id);
 		}
 	}
 }
